Ensure func_t & string_t have correct data types on 64 bit systems

+++ WinQuake/src/winquake/pr_comp.h
@@ -20,8 +20,10 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 // this file is shared by quake and qcc
 
-typedef int	func_t;
-typedef int	string_t;
+#include <cstdint>
+
+typedef intptr_t func_t;
+typedef intptr_t string_t;
 
 typedef enum : unsigned short {ev_void, ev_string, ev_float, ev_vector, ev_entity, ev_field, ev_function, ev_pointer} etype_t;
